requirements.txt
Flask>=3.0
Flask-WTF>=1.2
SQLAlchemy>=2.0
Flask-Migrate>=4.0
python-dotenv>=1.0
email-validator>=2.0
alembic==1.13.3
blinker==1.8.2
Bootstrap-Flask==2.3.3
click==8.1.7
colorama==0.4.6
Flask-CKEditor==0.5.1
Flask-Gravatar==0.5.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==3.0.2
packaging==24.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25
typing_extensions==4.12.2
Werkzeug==3.0.6
WTForms==3.0.1
Flask-Mail==0.9.1